requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.5.18.1
cloudinary==1.29.0
crispy-bootstrap5==0.6
dj-database-url==0.5.0
Django==4.0.4
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-registration-redux==1.3
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2022.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.2.0
